requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.5
apscheduler==3.10.1
python-dotenv==1.0.0
